moved __main__ to examples
