+++ indica/workflows/priors.py
@@ -268,21 +268,3 @@ def sample_from_high_density_region(
     return start_points
 
 
-@hydra.main(
-    version_base=None,
-    config_path="../configs/workflows/priors/",
-    config_name="config",
-)
-def main(cfg: DictConfig):
-    pm = PriorManager(**cfg)
-    post = pm.ln_prior({"electron_density.y0": 1e20, "electron_density.y1": 1e19})
-    samples = sample_from_priors(
-        ["electron_density.y0", "electron_density.y1"], pm.priors
-    )
-
-    print(post)
-    print(samples)
-
-
-if __name__ == "__main__":
-    main()
